requirements.txt
matplotlib==3.7.2
numba==0.58.1
numpy==1.24.4
onnx==1.14.0
onnxsim==0.4.33
opencv_contrib_python==4.8.1.78
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
Pillow==10.0.0
scipy==1.8.1
thop==0.1.1.post2209072238
torchsummary==1.5.1
tqdm==4.66.1